Newzoni merged

+++ src/main/java/doubleni/mealrecipe/controller/LikeController.java
@@ -18,6 +18,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 import static doubleni.mealrecipe.config.exception.BaseResponseStatus.*;
@@ -34,8 +35,10 @@ public class LikeController {
 
     @GetMapping("/all-board")
     @ApiOperation(value = "모든 게시판 좋아요 조회", notes = "Board 좋아요 레시피(리스트) 출력")
-    @ApiResponses(value={@ApiResponse(code =3000, message = "값을 불러오는데 실패하였습니다."),
-    @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요.")})
+    @ApiResponses(value={@ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 3000,message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code = 4000,message = "데이터베이스 연결에 실패하였습니다."),
+            @ApiResponse(code = 4053,message = "게시판 좋아요 조회 실패")})
     public BaseResponse<List<BoardLikeRes>> getAllBoardLike() {
         try {
             Long idx = jwtService.getUserIdx();
@@ -45,16 +48,18 @@ public class LikeController {
             List<BoardLikeRes> likes = likeService.getAllBoardLike();
             return new BaseResponse<>(likes);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(RESPONSE_ERROR);
         }
     }
 
     @GetMapping("/users-board")
     @ApiOperation(value = "사용자가 저장한 모든 게시판 조회", notes = "Board 좋아요 레시피(리스트) 출력")
-    @ApiResponses(value={@ApiResponse(code =3000, message = "값을 불러오는데 실패하였습니다."),
-            @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+    @ApiResponses(value={@ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
             @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
-            @ApiResponse(code = 2080, message = "좋아요가 존재 하지 않읍니다.")})
+            @ApiResponse(code = 3000, message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다."),
+            @ApiResponse(code = 4053, message = "게시판 좋아요 조회 실패")})
+
     public BaseResponse<List<BoardLikeRes>> getBoardLikeByUserId() throws BaseException {
         try {
             Long idx = jwtService.getUserIdx();
@@ -70,11 +75,14 @@ public class LikeController {
 
     @PostMapping("/board/add")
     @ApiOperation(value = "Board 좋아요(저장)", notes = "사용자가 Board를 좋아요(저장)한다.")
-    @ApiResponses(value={@ApiResponse(code =4000,message = "데이터베이스 연결에 실패하였습니다."),
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
             @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
-            @ApiResponse(code = 4050, message = "좋아요 추가 실패"),
-            @ApiResponse(code = 2081, message = "이미 좋아요한 게시글입니다.")})
-    public BaseResponse<BoardLikeRes> addBoardLike(@RequestParam("boardId") Long boardId) {
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 2061, message = "존재 하지 않거나 삭제된 게시글 입니다."),
+            @ApiResponse(code = 2080, message = "좋아요 실패"),
+            @ApiResponse(code = 2081, message = "이미 좋아요한 게시글입니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다.")})
+    public BaseResponse<BoardLikeRes> addBoardLike(@Valid @RequestParam("boardId") Long boardId) {
         try {
             Long idx = jwtService.getUserIdx();
 
@@ -82,10 +90,14 @@ public class LikeController {
                 return new BaseResponse<>(USERS_EMPTY_USER_ID);
             }
 
+            if (boardId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
+
             BoardLikeRes boardLikeRes = likeService.addBoardLike(idx, boardId);
             return new BaseResponse<>(boardLikeRes);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(ADD_FAIL_LIKE);
         }
     }
 
@@ -108,8 +120,12 @@ public class LikeController {
 
     @DeleteMapping("/board/delete")
     @ApiOperation(value = "Board 좋아요(저장) 삭제")
-    @ApiResponses(value={@ApiResponse(code =4000,message = "데이터베이스 연결에 실패하였습니다."),
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
             @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 2061, message = "존재 하지 않거나 삭제된 게시글 입니다."),
+            @ApiResponse(code = 2080, message = "좋아요가 존재 하지 않습니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다."),
             @ApiResponse(code = 4051, message = "좋아요 삭제 실패")})
     public BaseResponse<String> deleteBoardLike(@RequestParam("boardId") Long boardId) {
         try {
@@ -117,18 +133,48 @@ public class LikeController {
             if (idx == 0) {
                 return new BaseResponse<>(USERS_EMPTY_USER_ID);
             }
+            if (boardId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
             likeService.deleteBoardLike(boardId, idx);
             return new BaseResponse<>(boardId + "번 게시글 좋아요 취소!");
         } catch (BaseException e) {
-            return new BaseResponse<>(e.getStatus());
+            return new BaseResponse<>(DELETE_FAIL_LIKE);
+        }
+    }
+
+    @GetMapping("/check-board-like")
+    @ApiOperation(value = "사용자 Board 좋아요 확인", notes = "게시글 좋아요 여부 반환")
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
+            @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 2061, message = "존재 하지 않거나 삭제된 게시글 입니다."),
+            @ApiResponse(code = 2081, message = "좋아요가 존재 하지 않습니다."),
+            @ApiResponse(code = 3000, message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code =4000, message = "데이터베이스 연결에 실패하였습니다.")})
+    public BaseResponse<Long> checkBoardLike(@RequestParam("boardId") Long boardId) {
+        try {
+            Long idx = jwtService.getUserIdx();
+            if (idx == 0) {
+                return new BaseResponse<>(USERS_EMPTY_USER_ID);
+            }
+            if (boardId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
+            Long checkLike = likeService.checkBoardLike(idx, boardId);
+            return new BaseResponse<>(checkLike);
+        } catch (BaseException exception) {
+            return new BaseResponse<>(RESPONSE_ERROR);
         }
     }
 
     // ================================================================================
     @GetMapping("/all-recipe")
     @ApiOperation(value = "모든 레시피 좋아요 조회", notes = "사용자가 저장한 모든 레시피(리스트) 출력")
-    @ApiResponses(value={@ApiResponse(code =3000, message = "값을 불러오는데 실패하였습니다."),
-            @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요.")})
+    @ApiResponses(value={@ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 3000, message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다."),
+            @ApiResponse(code = 4054, message = "레시피 좋아요 조회 실패")})
     public BaseResponse<List<RecipeLikeRes>> getAllRecipeLike() {
         try {
             Long idx = jwtService.getUserIdx();
@@ -138,13 +184,16 @@ public class LikeController {
             List<RecipeLikeRes> likes = likeService.getAllRecipeLike();
             return new BaseResponse<>(likes);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(RESPONSE_ERROR);
         }
     }
     @GetMapping("/users-recipe")
     @ApiOperation(value = "사용자가 저장한 모든 레시피 조회", notes = "Recipe 좋아요 레시피(리스트) 출력")
-    @ApiResponses(value={@ApiResponse(code =3000, message = "값을 불러오는데 실패하였습니다."),
-            @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요.")})
+    @ApiResponses(value={@ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 3000,message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code = 4000,message = "데이터베이스 연결에 실패하였습니다."),
+            @ApiResponse(code = 4054,message = "레시피 좋아요 조회 실패")})
     public BaseResponse<List<RecipeLikeRes>> getRecipeLikeByUserId() {
         try {
             Long idx = jwtService.getUserIdx();
@@ -160,10 +209,13 @@ public class LikeController {
 
     @PostMapping("/recipe/add")
     @ApiOperation(value = "Recipe 좋아요(저장)")
-    @ApiResponses(value={@ApiResponse(code =4000,message = "데이터베이스 연결에 실패하였습니다."),
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
             @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
-            @ApiResponse(code = 4050, message = "좋아요 추가 실패"),
-            @ApiResponse(code = 2082, message = "이미 좋아요한 레시피입니다.")})
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 2050, message = "존재하지 않는 레시피입니다."),
+            @ApiResponse(code = 2080, message = "좋아요 실패"),
+            @ApiResponse(code = 2082, message = "이미 좋아요한 레시피입니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다.")})
     public BaseResponse<RecipeLikeRes> addRecipeLike(@RequestParam("rcpId") Long rcpId) {
         try {
             Long idx = jwtService.getUserIdx();
@@ -171,20 +223,25 @@ public class LikeController {
             if (idx == 0) {
                 return new BaseResponse<>(USERS_EMPTY_USER_ID);
             }
-
+            if (rcpId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
             RecipeLikeRes recipeLikeRes = likeService.addRecipeLike(idx, rcpId);
             return new BaseResponse<>(recipeLikeRes);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(ADD_FAIL_LIKE);
         }
     }
 
 
     @DeleteMapping("/recipe/delete")
     @ApiOperation(value = "Recipe 좋아요(저장) 삭제")
-    @ApiResponses(value={@ApiResponse(code =4000,message = "데이터베이스 연결에 실패하였습니다."),
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
             @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 2050, message = "존재하지 않는 레시피입니다."),
             @ApiResponse(code = 2080, message = "좋아요가 존재 하지 않읍니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다."),
             @ApiResponse(code = 4051, message = "좋아요 삭제 실패")})
     public BaseResponse<String> deleteRecipeLike(@RequestParam("rcpId") Long rcpId) {
         try {
@@ -192,29 +249,38 @@ public class LikeController {
             if (idx == 0) {
                 return new BaseResponse<>(USERS_EMPTY_USER_ID);
             }
+            if (rcpId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
             likeService.deleteRecipeLike(idx, rcpId);
             return new BaseResponse<>(rcpId + "번 레시피 좋아요 취소!");
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(DELETE_FAIL_LIKE);
         }
     }
 
     // rcpid값 넘겨서 좋아요 눌렀는지 여부 반환하는 api를 하나더 만드는거
     @GetMapping("/check-recipe-like")
     @ApiOperation(value = "사용자 Recipe 좋아요 확인", notes = "레시피 좋아요 여부 반환")
-    @ApiResponses(value={@ApiResponse(code =4000,message = "데이터베이스 연결에 실패하였습니다."),
+    @ApiResponses(value={@ApiResponse(code = 2000, message = "입력값을 확인해주세요."),
             @ApiResponse(code = 2010, message = "유저 아이디 값을 확인해주세요."),
-            @ApiResponse(code = 2080, message = "좋아요가 존재 하지 않읍니다.")})
+            @ApiResponse(code = 2011, message = "존재하지 않는 유저입니다."),
+            @ApiResponse(code = 2050, message = "존재하지 않는 레시피입니다."),
+            @ApiResponse(code = 3000, message = "값을 불러오는데 실패하였습니다."),
+            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다.")})
     public BaseResponse<Long> checkRecipeLike(@RequestParam("rcpId") Long rcpId) {
         try {
             Long idx = jwtService.getUserIdx();
             if (idx == 0) {
                 return new BaseResponse<>(USERS_EMPTY_USER_ID);
             }
+            if (rcpId == null) {
+                return new BaseResponse<>(REQUEST_ERROR);
+            }
             Long checkLike = likeService.checkRecipeLike(idx, rcpId);
             return new BaseResponse<>(checkLike);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(RESPONSE_ERROR);
         }
     }
 }
